refactor: split iOS runner-client into focused modules
Break the 1129-LOC runner-client.ts into cohesive modules:
- runner-errors.ts: error classification, hints, error builders
- runner-xctestrun.ts: xctestrun discovery, build, preparation, build settings
- runner-transport.ts: HTTP transport, endpoint resolution, connection waiting
- runner-session.ts: session lifecycle, process management, device validation
- runner-client.ts: thin orchestration + re-exports (preserves all public API)
